requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
pycryptodome
pytest
